Dev

See merge request itsmycargo/imc-react-api!782

Former-commit-id: f8342b5e8761c2e36d1603b6937b40dc9db962c7

+++ client/app/components/Admin/AdminShipmentView/AdminShipmentContent.jsx
@@ -101,31 +101,32 @@ class AdminShipmentContent extends Component {
 
       Object.keys(uploadedDocs).forEach((key) => {
         docChecker[key] = true
-
-        docView.push(<div className={`flex-35 layout-row layout-align-start-start layout-padding ${adminStyles.uploaded_doc}`}>
-          <i className="fa fa-check flex-none" style={{ color: 'rgb(13, 177, 75)' }} />
-          <div className="layout-row flex layout-wrap" style={{ marginBottom: '12px' }}>
-            <h4 className="flex-100 layout-row">{documentTypes[key]}</h4>
-            {uploadedDocs[key].map(doc => (
-              <div className="flex-100 layout-row">
-                <a
-                  href={doc.signed_url}
-                  className={`${styles.eye_link} flex-none layout-row layout-align-center-center`}
-                  target="_blank"
-                >
-                  <i className="fa fa-eye pointy flex-none" />
-                </a>
-                <i
-                  className="fa fa-trash pointy flex-none"
-                  onClick={() => this.deleteDoc(doc)}
-                />
-                <p className="flex layout-row">
-                  {doc.text}
-                </p>
-              </div>
-            ))}
-          </div>
-        </div>)
+        if (key !== 'shipment_recap') {
+          docView.push(<div className={`flex-35 layout-row layout-align-start-start layout-padding ${adminStyles.uploaded_doc}`}>
+            <i className="fa fa-check flex-none" style={{ color: 'rgb(13, 177, 75)' }} />
+            <div className="layout-row flex layout-wrap" style={{ marginBottom: '12px' }}>
+              <h4 className="flex-100 layout-row">{documentTypes[key]}</h4>
+              {uploadedDocs[key].map(doc => (
+                <div className="flex-100 layout-row">
+                  <a
+                    href={doc.signed_url}
+                    className={`${styles.eye_link} flex-none layout-row layout-align-center-center`}
+                    target="_blank"
+                  >
+                    <i className="fa fa-eye pointy flex-none" />
+                  </a>
+                  <i
+                    className="fa fa-trash pointy flex-none"
+                    onClick={() => this.deleteDoc(doc)}
+                  />
+                  <p className="flex layout-row">
+                    {doc.text}
+                  </p>
+                </div>
+              ))}
+            </div>
+          </div>)
+        }
       })
     }
     Object.keys(docChecker).forEach((key) => {
@@ -316,7 +317,7 @@ class AdminShipmentContent extends Component {
                               style={shipment.has_pre_carriage ? selectedStyle : deselectedStyle}
                             />
                             <p>
-                              {t('shipment:originDocumentation')}
+                              {t('shipment:originLocalCharges')}
                             </p>
                           </div>
                           {feeHash.export ? <div className="flex-100 layout-row layout-align-end-center">
@@ -363,7 +364,7 @@ class AdminShipmentContent extends Component {
                               style={shipment.has_on_carriage ? selectedStyle : deselectedStyle}
                             />
                             <p>
-                              {t('shipment:destinationDocumentation')}
+                              {t('shipment:destinationLocalCharges')}
                             </p>
                           </div>
                           {feeHash.import ? <div className="flex-100 layout-row layout-align-end-center">
